Restrict filenames more granularly (READ MSG)
Windows has a MAX_PATH length of 260 characters which some video titles exceed when combined with the path to the download folder. AYC does not have any control over where the user decides to set the download directory.
So all filenames are limited to 128 (+4 or +5 for the file extension) characters. Leaving the remaining 127-128 characters (depending on the file extension) for the directory path.
In this commit, the 128 characters for the file name are divided among various parts of the filename in an optimal way to ensure functionality and user friendliness.

If you are already at a limit and want to set AYC Download directory beyond the 127-128 limit follow this guide to enable long path names: https://www.autodesk.com/support/technical/article/caas/sfdcarticles/sfdcarticles/The-Windows-10-default-path-length-limitation-MAX-PATH-is-256-characters.html
In case the above URL does not work anymore, simply google "Remove Windows Max Path Limit".